- remove UseFormsAppCompat extension
