+++ src/Compatibility/Core/src/Compatibility.csproj
@@ -60,7 +60,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <Compile Include="AppHostBuilderExtensions.Controls.cs" />
     <Compile Include="AppHostBuilderExtensions.cs" />
     <Compile Include="MauiHandlersCollectionExtensions.cs" />
     <Compile Include="Forms.cs" />
